requirements.txt
https://github.com/bakfu/bakfu.git
functools32
pyyaml
six
jsonpickle
